requirements.txt
gitpython>=3.1.0
mcp[cli]>=1.0.0
